Move matplotlib_svg_scraper in a new file named utils.py in the folder docs
