+++ docs/conf.py
@@ -34,6 +34,10 @@ rtd_version = os.environ.get('READTHEDOCS_VERSION', 'local')
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
+
+# Where our helpers live
+util_root = "tslearn.docs.utils"
+
 extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.autosummary',
@@ -61,18 +65,6 @@ intersphinx_mapping = {
     'sklearn': ('https://scikit-learn.org/stable', None)
 }
 
-from sphinx_gallery.scrapers import matplotlib_scraper
-
-class MatplotlibSVGScraper(object):
-
-    def __repr__(self):
-        return self.__class__.__name__
-
-    def __call__(self, *args, **kwargs):
-        return matplotlib_scraper(*args, format='svg', **kwargs)
-
-matplotlib_svg_scraper = MatplotlibSVGScraper()
-
 sphinx_gallery_conf = {
     'examples_dirs': ['./examples'],
     'gallery_dirs': ['./auto_examples'],
@@ -83,7 +75,7 @@ sphinx_gallery_conf = {
     'subsection_order': ["examples", "examples/metrics", "examples/neighbors",
                          "examples/clustering", "examples/classification",
                          "examples/autodiff", "examples/misc"].index,
-    'image_scrapers': ("matplotlib_svg_scraper",),
+    'image_scrapers': (f"{util_root}.matplotlib_svg_scraper",),
     # 'binder': {
     #     # Required keys
     #     'org': 'rtavenar',
